% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indices.R
\name{computeAnnualEfficiency}
\alias{computeAnnualEfficiency}
\title{Compuate Efficiency for Each Year}
\usage{
computeAnnualEfficiency(sim, resource, world)
}
\description{
Compuate Efficiency for Each Year
}
