% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indices.R
\name{computeAvgTotalError}
\alias{computeAvgTotalError}
\title{Computing Average Total Mismatch}
\usage{
computeAvgTotalError(sim, world, par0)
}
\arguments{
\item{sim}{A simulation run from runManyYears.}

\item{world}{world object; list of 7: a population distribution across the time period in a T x X matrix,
 a vector with midpoint X-values, the time points for the population as integers 1:tau, the minimum value of population distribution
 (X.min), the maximum value of population distribution (X.max),
 the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 8.
 Set up by the getSinePop/getOptimal function. 
 @param par0 Data frame containing resource peak timing and space information of x.peak, t.peak, x.sd
and t.sd for each year Generated by the function getCCpars
 @return a single average total mismatch (avgTE) index
 @seealso \link{getCCpars}, \link{computeTotalError}, \link{computeMigrationIndices}
 @example examples/indices_examples.R}
}
\description{
Computes the average total mismatch (TE) of a migratory population and a drifting/stochastic resource
by computing the average of the mismatch for each population year with its respective resource 
for the last 20 years of a simulation run.
}
