% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indices.R
\name{computeEfficiency}
\alias{computeEfficiency}
\title{Compute Efficiency}
\usage{
computeEfficiency(pop, resource, world)
}
\arguments{
\item{pop}{One cycle of population process}

\item{resource}{The resource in an given cycle to compare to pop}

\item{world}{world object}
}
\value{
a single foraging efficiecy (FE) index
}
\description{
Returns an efficiency of foraging index of a single cycle population 
process, based on comparing the population distribution to a resource
distribution.
}
