% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indices.R
\name{computeSpatialAdaptationIndex}
\alias{computeSpatialAdaptationIndex}
\title{Keeping up with Climate Change - Spatial Adaptation Index}
\usage{
computeSpatialAdaptationIndex(sim, cc.params, trim = 1)
}
\description{
Keeping up with Climate Change - Spatial Adaptation Index
}
