% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indices.R
\name{computeTemporalAdaptationIndex}
\alias{computeTemporalAdaptationIndex}
\title{Keeping up with Climate Change - Temporal Adaptation Index}
\usage{
computeTemporalAdaptationIndex(sim, cc.params, trim = 1)
}
\arguments{
\item{sim}{A simulation run from runManyYears.}

\item{cc.params}{Data frame containing resource peak timing and space information of x.peak, t.peak, x.sd
and t.sd. Generated by the function getCCpars}
}
\value{
a single temporal adaptation (TA) index
}
\description{
Computes the temporal adaptation (TA) index which is the slope of the migration timing parameter
against the resource drift
}
\seealso{
\link{getCCpars}
}
