% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateScripts.R
\name{createShellScript}
\alias{createShellScript}
\title{Create Shell Script}
\usage{
createShellScript(shell.dir, code.dir, runname, filename)
}
\arguments{
\item{shell.dir}{...}

\item{code.dir}{...}

\item{runname}{...}

\item{filename}{string containing base name of the file of the R.script to run 
through the shell}
}
\value{
creates .sh files
}
\description{
Generates .sh files to run many different parameters of the same model through
the shell.
}
\seealso{
\link{createSource}, \link{parameterGrid}
}
