% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateScripts.R
\name{createShellScripts}
\alias{createShellScripts}
\title{Create Shell Scripts}
\usage{
createShellScripts(shell.dir, code.dir, runname, filename)
}
\arguments{
\item{shell.dir}{string containing directory of where shell scripts will be created}

\item{code.dir}{string containing directory of where R scripts our scored}

\item{runname}{string containing run name for server}

\item{filename}{string containing base name of the file of the R.script to run 
through the shell}
}
\value{
creates a .sh file for each R script in code.dir
}
\description{
Generates .sh files to run many different parameters of the same model through
the shell.
}
\seealso{
\link{createSource}, \link{parameterGrid}, \link{createFinalShellScript}
}
