% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteMigrationFunctions.R
\name{fitMigration}
\alias{fitMigration}
\title{Fit Migration}
\usage{
fitMigration(
  t,
  x,
  m.start = NULL,
  tau = 100,
  lower = c(-100, 0, 0, 0, -100, -100)
)
}
\description{
Fit Migration
}
