% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForagingMemoryModel.R
\name{ForagingMemoryModel}
\alias{ForagingMemoryModel}
\title{Foraging with memory model}
\usage{
ForagingMemoryModel(t, pop, parms, memory, resource, dx = dx)
}
\arguments{
\item{t}{vector of times}

\item{pop}{a T x X matrix describing the population distribution 
across the time period.}

\item{parms}{named vector of parameters. These are \code{epsilon} - diffusion coefficient; 
\code{alpha} - resource 
following coefficient; \code{beta0} - social cohesion coefficient; \code{beta1} - 
memory coefficient}

\item{memory}{the memory score}

\item{resource}{a T x X distribution of the (potentially) dynamic resource.}

\item{dx}{time step}
}
\value{
Model setup; output of \code{\link{tran.1D}} function.
}
\description{
This function sets up the diffusion advection model
}
