% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPulsedResource.R
\name{getCCpars}
\alias{getCCpars}
\title{Get Climate Parameters}
\usage{
getCCpars(
  mu_x0,
  mu_t0,
  sigma_x,
  sigma_t,
  n.years,
  beta_x = 0,
  beta_t = 0,
  psi_x = 0,
  psi_t = 0,
  n.years.null = 0
)
}
\arguments{
\item{mu_x0}{spatial coordinate of resource peark for summer}

\item{mu_t0}{timing of 
resource peak for the summer}

\item{sigma_x}{spatial scale of resource pulse}

\item{sigma_t}{time duration of resource pulse}

\item{n.years}{how many years the resource should be created}

\item{beta_x}{rate of change of peak location}

\item{beta_t}{rate of change of peak timing}

\item{psi_x}{standard deviation of peak location}

\item{psi_t}{standard deviation of peak timing}

\item{n.years.null}{how many years the resource remains stable i.e. no drift or stochasticity}
}
\value{
A data frame containing values of t.peak (which is the maximum peak of t), 
t.sd (which is the standard deviation of t), x.peak (which is the maximum peak of x),
and x.sd (which is the standard deviation of x)
}
\description{
Creates a data frame of resource parameters to create an island or drifigng resource
}
\examples{
## Worlds
  world_nonmigratory <- getSinePop(tau = 100, peak.max = 1, peak.min = -1, sd = 10)
  world_sinusoidal <- getSinePop(tau = 100, peak.max = 40, peak.min = -40, sd = 10)
  world_optimal <- getOptimalPop(tau = 100, t.peak = 25, 
                                 x1 = 40, x2 = -40, x.sd = 5, t.sd = 12)
  
  
  require(fields)
  with(world_nonmigratory, image.plot(time, X, pop))
  with(world_sinusoidal, image.plot(time, X, pop))
  with(world_optimal, image.plot(time, X, pop))
  
  
  
## Fixed (bimodal) resource example

  resource_drifting <- getResource_drifting(world_sinusoidal, 
                                         par = c(t.peak = 25, t.sd = 12, 
                                                 x.peak = 30, x.sd = 6))
  
  with(world_sinusoidal, image.plot(time, X, resource_drifting))
  
  
## Independent (Island) resource
  
  resource_island <- getResource_island(world_sinusoidal, 
                                            par = c(t.peak = 25, t.sd = 6, 
                                                    x.peak = 30, x.sd = 12))
  
  with(world_sinusoidal, image.plot(time, X, resource_island))
  
  
## Climate Change Island Resource
  par0 <- getCCpars(mu_x0 = 30, 
                    mu_t0 = 25,
                    beta_x = -1,
                    beta_t = 0,
                    n.years = 5,
                    sigma_x = 12,
                    sigma_t = 6,
                    psi_x = 0, 
                    psi_t = 0,
                    n.years.null = 0)
  world$resource <- aaply(par0, 1, function(p) getResource_island(world, p))

  
  
}
\seealso{
\link{getResource_island}, \link{getResource_drifting}
}
