% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilingAndPlotting.R
\name{getCompiledResults}
\alias{getCompiledResults}
\title{Get Compiled Results}
\usage{
getCompiledResults(results.dir)
}
\arguments{
\item{results.dir}{string containing directory of where results are stored}
}
\value{
creates a data frame of all the results
}
\description{
Compiles results from different folders into one data frame
}
