% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialPop.R
\name{getOptimalPop}
\alias{getOptimalPop}
\title{Get Optimal Pop
Initilaizes optimal initial population distribution based on resource}
\usage{
getOptimalPop(
  tau,
  X.min = -100,
  X.max = 100,
  dx = 1,
  x1,
  x2,
  t.peak,
  x.sd,
  t.sd
)
}
\arguments{
\item{tau}{maximum time in days for migratory population in one year}

\item{X.min}{minimum value of population distribution}

\item{X.max}{maximum value of population distribution}

\item{dx}{time step}

\item{t.peak}{where the resource peak is in time}

\item{x.sd}{standard deviation of resource peak in space}

\item{t.sd}{standard deviation of resource peak in time}

\item{x.peak}{where the resource peak is in space}
}
\description{
Get Optimal Pop
Initilaizes optimal initial population distribution based on resource
}
