% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPulsedResource.R
\name{getResource_drifting}
\alias{getResource_drifting}
\title{Get Resource Drifting}
\usage{
getResource_drifting(world, par, x.null = 0)
}
\arguments{
\item{world}{world object (e.g. from getSinePop/getOptimalPop); list of 7: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau, the minimum value of population distribution
(X.min), the maximum value of population distribution (X.max),
the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 8.}

\item{par}{vector with four values: t.peak (which is the maximum peak of t), 
t.sd (which is the standard deviation of t), x.peak (which is the maximum peak of x),
and x.sd (which is the standard deviation of x).}

\item{x.null}{extent of "dead area" - needed to keep the process from extending too far beyond the limits of the range.}
}
\value{
A X x X matrix containing values of the resource distribution
}
\description{
This function generates a seasonal resource function with the following properties:
1. The total amount of resource across space is constant throughout the year.
2. At the beginning, middle, and end of the year the resource is uniformly distributed. 
3. At some peak time $t_r < tau/2$, the resource concentrates at a location $x_r < chi/2$
 with a spatial deviation $sigma_x$ and a temporal deviation $sigma_t$ (where $tau$ 
 is the length of the year and $chi$ is the extent of the spatial domain).
4. The resource peaks exactly symmetrically at time $tau - t_r$ and location 
$chi - x_r$ with the same variance $sigma_r$.
}
\details{
To generate a resource with these properties, we distributing the resource in 
space as a beta distribution, where the two shape and scale parameters vary 
sinusoidally in such a way as to fulfill the criteria above.
}
\examples{
## Worlds
  world_nonmigratory <- getSinePop(tau = 100, peak.max = 0, peak.min = 0, sd = 10)
  world_sinusoidal <- getSinePop(tau = 100, peak.max = 40, peak.min = -40, sd = 10)
  world_optimal <- getOptimalPop(tau = 100, t.peak = 25, 
                                 x1 = 40, x2 = -40, x.sd = 5, t.sd = 12)
  
  
  require(fields)
  with(world_nonmigratory, image.plot(time, X, pop))
  with(world_sinusoidal, image.plot(time, X, pop))
  with(world_optimal, image.plot(time, X, pop))
  
  
  
## Fixed (bimodal) resource example

  resource_drifting <- getResource_drifting(world_sinusoidal, 
                                         par = c(t.peak = 25, t.sd = 12, 
                                                 x.peak = 40, x.sd = 5))
  
  with(world_sinusoidal, image.plot(time, X, resource_drifting))
  
  
## Independent (Island) resource
  
  resource_island <- getResource_island(world_sinusoidal, 
                                            par = c(t.peak = 25, t.sd = 12, 
                                                    x.peak = 80, x.sd = 5))
  
  with(world_sinusoidal, image.plot(time, X, resource_island))
  
  
## Climate Change Island Resource
  par0 <- getCCpars(mu_x0 = 80, 
                    mu_t0 = 25,
                    beta_x = -1,
                    beta_t = 0,
                    n.years = 5,
                    sigma_x = 5,
                    sigma_t = 12,
                    psi_x = 0, 
                    psi_t = 0)
  world$resource <- aaply(par0, 1, function(p) getResource_island(world, p))

  
  
}
\seealso{
\link{getResource_island}, \link{getCCpars}
}
