% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPulsedResource.R
\name{getResource_island}
\alias{getResource_island}
\title{#' Get Island}
\usage{
getResource_island(world, par)
}
\arguments{
\item{world}{world object (e.g. from getSinePop/getOptimalPop); list of 7: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau, the minimum value of population distribution
(X.min), the maximum value of population distribution (X.max),
the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 8.}

\item{par}{vector with four values: t.peak (which is the maximum peak of t), 
t.sd (which is the standard deviation of t), x.peak (which is the maximum peak of x),
and x.sd (which is the standard deviation of x). can be created by getCCpars for multiple years}
}
\value{
A X x X matrix containing values of the resource distribution
}
\description{
This function generates a seasonal pulsed resource function.
}
\examples{
## Worlds
  world_nonmigratory <- getSinePop(tau = 100, peak.max = 0, peak.min = 0, sd = 10)
  world_sinusoidal <- getSinePop(tau = 100, peak.max = 40, peak.min = -40, sd = 10)
  world_optimal <- getOptimalPop(tau = 100, t.peak = 25, 
                                 x1 = 40, x2 = -40, x.sd = 5, t.sd = 12)
  
  
  require(fields)
  with(world_nonmigratory, image.plot(time, X, pop))
  with(world_sinusoidal, image.plot(time, X, pop))
  with(world_optimal, image.plot(time, X, pop))
  
  
  
## Fixed (bimodal) resource example

  resource_drifting <- getResource_drifting(world_sinusoidal, 
                                         par = c(t.peak = 25, t.sd = 12, 
                                                 x.peak = 40, x.sd = 5))
  
  with(world_sinusoidal, image.plot(time, X, resource_drifting))
  
  
## Independent (Island) resource
  
  resource_island <- getResource_island(world_sinusoidal, 
                                            par = c(t.peak = 25, t.sd = 12, 
                                                    x.peak = 80, x.sd = 5))
  
  with(world_sinusoidal, image.plot(time, X, resource_island))
  
  
## Climate Change Island Resource
  par0 <- getCCpars(mu_x0 = 80, 
                    mu_t0 = 25,
                    beta_x = -1,
                    beta_t = 0,
                    n.years = 5,
                    sigma_x = 5,
                    sigma_t = 12,
                    psi_x = 0, 
                    psi_t = 0)
  world$resource <- aaply(par0, 1, function(p) getResource_island(world, p))

  
  
}
\seealso{
\link{getResource_drifting}, \link{getCCpars}
}
