% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorld.R
\name{getSinePop}
\alias{getSinePop}
\title{Get Sine Pop}
\usage{
getSinePop(tau, X.min = -100, X.max = 100, dx = 1, peak.max, peak.min, sd)
}
\arguments{
\item{tau}{maximum time in days for migratory population in one year}

\item{X.min}{minimum value of population distribution}

\item{X.max}{maximum value of population distribution}

\item{dx}{time step}

\item{peak.max}{minimum value of sinusoidal peak}

\item{peak.min}{maximum value of sinusoidal peak}

\item{sd}{standard deviation of population distribution}
}
\value{
list of 5 which includes a population distribution across the time period in a T x X matrix,
 a vector with midpoint X-values, the time points for the population as integers 1:tau,
 the dx value and the tau value.
 @example examples/WorldsandResources.R
}
\description{
Initializes sinusoidal year 0 of a migratory population distribution 
via a sinusoidal migration pattern
}
