% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateScripts.R
\name{parameterGrid}
\alias{parameterGrid}
\title{Parameter Grid}
\usage{
parameterGrid(epsilons, alphas, betas, kappas, lambdas, existing = NULL)
}
\arguments{
\item{epsilons}{values of epsilon parameter in a vector}

\item{alphas}{values of alpha parameter in a vector}

\item{betas}{values of beta parameter in a vector}

\item{kappas}{values of kappa parameter in a vector}

\item{lambdas}{values of lambda parameter in a vector}

\item{existing}{a data frame with existing values already tested}
}
\value{
list of data frames
}
\description{
Generates a data frame splitting up all of the combinations of parameters
}
\examples{
parameterGrid(c(1,2,3), c(2,3), seq(1,4), 4, 20)
}
\seealso{
\link{createShellScript}, \link{createSource}, \link{createFinalShellScript}
}
