% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateScripts.R
\name{parameterGridres}
\alias{parameterGridres}
\title{Parameter Grid for Resource}
\usage{
parameterGridres(params, existing)
}
\description{
Generates a data frame splitting up all of the combinations of parameters and considers resource
}
