% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilingAndPlotting.R
\name{plotCompiled}
\alias{plotCompiled}
\title{Plot Compiled}
\usage{
plotCompiled(r)
}
\arguments{
\item{r}{data to be plotted}
}
\value{
three grid plots for parameters of Social Cohesion, Foraging Efficiency, 
and Migratoriness
}
\description{
Plots compiled results
}
