% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotManyRuns}
\alias{plotManyRuns}
\title{Plotting simulation results}
\usage{
plotManyRuns(
  sim,
  world,
  years = NULL,
  nrow = 1,
  outer = TRUE,
  labelyears = FALSE,
  par = NULL,
  ylab = "",
  ...
)
}
\arguments{
\item{sim}{A simulation run from runManyYears. Use sim$pop for the population.}

\item{world}{world object; list of 7: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau, the minimum value of population distribution
(X.min), the maximum value of population distribution (X.max),
the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 8.
Set up by the getSinePop/getOptimal function. 
@example examples/plotting_examples.R}
}
\description{
This function plots the migratoriness population distribution over a time period.
}
