% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotMigration}
\alias{plotMigration}
\title{Plot migration patterns over resource}
\usage{
plotMigration(M, world, plotresource = TRUE, add = FALSE)
}
\arguments{
\item{M}{A simulation run from runManyYears.}

\item{world}{world object; list of 7: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau, the minimum value of population distribution
(X.min), the maximum value of population distribution (X.max),
the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 8.
Set up by the getSinePop/getOptimal function.
@return a plot of the average migration of a simulation run for each year}
}
\description{
Plot migration patterns over resource
}
