% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotMigrationHat}
\alias{plotMigrationHat}
\title{Plot estimated migration (with or without memory)}
\usage{
plotMigrationHat(
  mhat,
  x.peak = NULL,
  t.peak = NULL,
  cols = c("darkorange", "darkblue"),
  legend = TRUE
)
}
\arguments{
\item{mhat}{data frame of migration estimates}

\item{{x.peak, t.peak}}{}
}
\description{
Plot estimated migration (with or without memory)
}
