% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotYearList}
\alias{plotYearList}
\title{Plot Year List}
\usage{
plotYearList(yearlist, X.max = 100, tau = 360, log = FALSE, persp = FALSE, ...)
}
\arguments{
\item{yearlist}{a list of T x X matrices describing the population 
distribution for each year after initial population; produced by the runManyYear function}

\item{X.max}{maximum value of X-axis}

\item{tau}{maximum value of time T}

\item{log}{logical;}

\item{persp}{whether to plot an image / contour or perspective plot}
}
\description{
This function plots the migratoriness population distribution over a time period.
}
\examples{
world <- getSinePop(tau = 100, X.min = 0, X.max = 100, dx=1, 
                    peak.max = 80, peak.min = 20, sd = 10)
par(mfrow = c(1,2))
plotYearList(list(world$pop))
plotYearList(list(world$pop), persp = TRUE)
}
\seealso{
\link{runManyYears}; \link{printParameters}
}
