% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{printParameters}
\alias{printParameters}
\title{Print Parameters}
\usage{
printParameters(p)
}
\description{
This function prints the parameters of the diffusion-advection equation used 
on the model on the plot of the population distribution set up by the plotYearList function.
}
\details{
@param p List of 5 a population distribution across the time period in a T x X matrix,
 a vector with midpoint X-values, the time points for the population as integers 1:tau,
 the dx value and the tau value. Set up by the getSinePop function. 
 @return displays the values of the parameters on the plot which are:
  \code{epsilon} - diffusion coefficient; \code{alpha} - resource 
following coefficient; \code{beta0} - social cohesion coefficient; \code{beta1} - 
memory coefficient
 @seealso \link{PlotYearList}; \link{getSinePop}
}
