% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModel.R
\name{runManyRunsmig}
\alias{runManyRunsmig}
\title{Run Many Runs Mig}
\usage{
runManyRunsmig(
  parameters.df,
  resource_param,
  world,
  resource,
  filename = NULL,
  results.dir = NULL,
  ...
)
}
\description{
Running on server for initial population constant
}
