% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModel.R
\name{runManyYears}
\alias{runManyYears}
\title{Run Many Years}
\usage{
runManyYears(
  World,
  Parameters,
  n.years = 30,
  threshold = 0.995,
  verbose = FALSE
)
}
\arguments{
\item{World}{World List of 5: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau,
the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 6.
Set up by the getSinePop function.}

\item{Parameters}{named vector of parameters. These are \code{epsilon} - diffusion coefficient; 
\code{alpha} - resource 
following coefficient; \code{beta0} - social cohesion coefficient; \code{beta1} - 
memory coefficient}

\item{n.years}{number of years the population migrates}

\item{threshold}{the threshold for the social cohesion between two population 
distributions from two consecutive years. This is a number between 0 and 1.}
}
\value{
a list of n.years containing T x X matrices describing the population 
distribution for each year after initial population
}
\description{
Based on a migratory population's set up (the World) and the values \code{alpha},
\code{beta0} and \code{beta1}, this function determines the population 
distribution after several years.
}
\seealso{
\link{getSinePop}, \link{runManyYears}
}
