% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModel.R
\name{runManyYearsWithStabilization}
\alias{runManyYearsWithStabilization}
\title{Run Many Years with Stabilization}
\usage{
runManyYearsWithStabilization(
  world,
  parameters,
  n.years = dim(world$resource)[1],
  threshold = threshold,
  verbose = FALSE,
  m0 = world$m0,
  n.years.null = 20,
  resource.null
)
}
\arguments{
\item{n.years}{number of years the population migrates}

\item{threshold}{the threshold for the social cohesion between two population 
distributions from two consecutive years. This is a number between 0 and 1.}

\item{m0}{the inital migration estimates for the population; created by fitMigration function}

\item{n.years.null}{number of years the population should stabilize for}

\item{resource.null}{Data frame containing the stable resource peak timing and space
information of x.peak, t.peak, x.sd and t.sd}

\item{World}{world object; list of 7: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau, the minimum value of population distribution
(X.min), the maximum value of population distribution (X.max),
the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 8.
Set up by the getSinePop/getOptimal function.}

\item{Parameters}{named vector of parameters. These are \code{epsilon} - diffusion coefficient; 
\code{alpha} - forage following coefficient; \code{beta} - strength of sociality; \code{lambda} - 
spatial sclae of sociality; \code{kappa} - propotion of reference versus working memory}
}
\value{
a list of n.years containing T x X matrices describing the population 
distribution for each year after initial population
}
\description{
A simulation run to reach a until a quasi-equilibrium
state was achieved, i.e. where the Bhattacharya index of the population distribution
across subsequent years reached a value of 1-1e-5
}
\seealso{
\link{getSinePop},\link{getOptimalPop}, \link{runNextYear}, \link{runManyYears}, \link{fitMigration}
}
