% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNextYear.R
\name{runNextYear}
\alias{runNextYear}
\title{Run Next Year}
\usage{
runNextYear(world, Parameters, Pop_lastyear, Year)
}
\arguments{
\item{Parameters}{named vector of parameters. These are \code{epsilon} - diffusion coefficient; 
\code{alpha} - resource following coefficient; \code{beta} - spatial scale of sociality; \code{kappa} - 
memory following coefficient; \code{lambda} - maximum speed}

\item{World}{world object; list of 7: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau, the minimum value of population distribution
(X.min), the maximum value of population distribution (X.max),
the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 8.
Set up by the getSinePop/getOptimal function.}
}
\value{
a T x X matrix describing the population distribution for the next year
}
\description{
Based on a migratory population's set up (the World) and the values \code{epsilon}, 
\code{alpha}, \code{beta}, \code{kappa} and \code{lambda}, this function determines the population 
distribution for the following year.
}
\seealso{
\link{getSinePop}, \link{getOptimalPop}, \link{runManyYears}
}
