% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteMigrationFunctions.R
\name{stepMigration}
\alias{stepMigration}
\title{Discrete migration functions}
\usage{
stepMigration(t, t1, dt1, t2, dt2, x1, x2, tau = 100)
}
\description{
Functions to fit the discrete step-wise migration process
}
