% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indices.R
\name{computeCohesiveness}
\alias{computeCohesiveness}
\title{Compute social cohesiveness}
\usage{
computeCohesiveness(pop, world)
}
\arguments{
\item{pop}{One cycle of population process}

\item{world}{world object; list of 5: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau,
the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 6.
Set up by the getSinePop function.}
}
\value{
a names vector of mean and standard deviation of SC
}
\description{
Returns a cohesiveness index of a single cycle population 
process: CI = 1 - MSD(X) / SD_max, where 
SD_max = X_max/sqrt(12), i.e. the standard deviation of a 
uniform distribution in the domain 0-X_max; and MSD(X) is 
the mean standard deviation of the population.
}
\examples{
# Initialize world

dx <- 2

parameters <- c(epsilon = 5, alpha = 300, beta0=100, beta1 = 300)
world.R1 <- getSinePop(tau = 100, X.min = 0, X.max = 100, dx=2, 
                    peak.max = 80, peak.min = 20, sd = 10)
world.R1$resource <- getPulsedResource(world.R1$time, world.R1$X, c(t.peak = 25, t.sd = 9, 
                                                           x.peak = 80, x.sd = 9))

#run simulation for 12 years 
M <- runManyYears(world.R1, Parameters = parameters, n.years = 6, 1)

world.R2 <- world.R1 \%>\% list_modify(pop = M$Year6)
M2 <- runManyYears(world.R2, Parameters = parameters, n.years = 6, 1)
plotYearList(M2, world.R2, tau = tau)

#simulation runs for 3 years since that is when threshold is hit

M <- runManyYears(world.R1, Parameters = parameters, n.years = 6, 0.99)
plotYearList(M, world.R1, tau = tau)

computeCohesiveness(M[[6]], world.R1)["SC.mean"]
computeMigratoriness(M[[6]], world.R1)$overlap
computeEfficiency(M[[6]], world.R1$resource, world.R1)
computeIndices(M[[6]], world.R1$resource, world.R1)
}
\seealso{
\link{computeEfficiency}, \link{computeMigratoriness}, \link{computeIndices}
}
