% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indices.R
\name{computeSpatialAdaptationIndex}
\alias{computeSpatialAdaptationIndex}
\title{Keeping up with Climate Change - Spatial Adaptation Index}
\usage{
computeSpatialAdaptationIndex(sim, cc.params, trim = 1)
}
\arguments{
\item{sim}{A simulation run from runManyYears.}

\item{cc.params}{Data frame containing resource peak timing and space information of x.peak, t.peak, x.sd
and t.sd. Generated by the function getCCpars}
}
\value{
a single spatial adaptation (SA) index
}
\description{
Computes the spatial adaptation (SA) index which is the slope of the migration location parameter
against the resource drift
}
\seealso{
\link{getCCpars}
}
