% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_indices.R
\name{computeTemporalAdaptationIndex}
\alias{computeTemporalAdaptationIndex}
\title{Keeping up with Climate Change - Temporal Adaptation Index}
\usage{
computeTemporalAdaptationIndex(sim, cc.params, trim = 1)
}
\description{
Keeping up with Climate Change - Temporal Adaptation Index
}
