% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateScripts.R
\name{createFinalShellScript}
\alias{createFinalShellScript}
\title{Create Final Shell Scripts}
\usage{
createFinalShellScript(shell.dir, runname)
}
\arguments{
\item{shell.dir}{string containing directory of where R scripts will be created}

\item{runname}{string containing run name for server}
}
\value{
creates one .sh file
}
\description{
Generates one .sh file listing all of the .sh files for model to run at once
}
\seealso{
\link{createSource}, \link{parameterGrid}, \link{createShellScripts}
}
