% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateScripts.R
\name{createSource}
\alias{createSource}
\title{Create Sources}
\usage{
createSource(
  worldname = "world",
  resourcename,
  code.dir,
  filename,
  results.dir,
  epsilons,
  alphas,
  beta0s,
  beta1s,
  existing,
  existingfile
)
}
\arguments{
\item{worldname}{string containing name of the world}

\item{resourcename}{string containing name of resource}

\item{code.dir}{string containing directory of where R scripts will be created}

\item{filename}{string containing base name of the file to be created}

\item{results.dir}{string containing directory of where results from the R scripts will be stored}

\item{epsilons}{values of epsilon parameter}

\item{alphas}{values value of alpha parameter}

\item{beta0s}{values value of beta0 parameter}

\item{beta1s}{values value of beta1 parameter}

\item{existing}{a data frame with existing values already tested
for server to access;  if there is no existing values, enter an empty data frame. 
Save this data frame as results.}

\item{existingfile}{name of file containing a data frame with existing values already tested
for server to access; 
if there is no existing values, enter create a file with an empty data frame}
}
\value{
creates .R files with scripts
}
\description{
Generates R.scripts to run many different parameters of the same model
}
\examples{

}
\seealso{
\link{createShellScript}, \link{parameterGrid}, \link{createFinalShellScript}
}
