% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForagingMemoryModel.R
\name{ForagingWithMemory}
\alias{ForagingWithMemory}
\title{Foraging with memory model}
\usage{
ForagingWithMemory(t, pop, parms, v.memory, dh, dp, dx = dx)
}
\arguments{
\item{t}{vector of times}

\item{pop}{a T x X matrix describing the population distribution 
across the time period.}

\item{parms}{named vector of parameters. These are \code{epsilon} - diffusion coefficient; 
\code{alpha} - resource following coefficient; \code{beta} - spatial scale of sociality; \code{kappa} - 
memory following coefficient; \code{lambda} - maximum speed}

\item{dh}{a row of a resource matrix (a T x X distribution of the (potentially) dynamic resource.)}

\item{dx}{time step}

\item{memory}{the memory score}
}
\value{
Model setup; output of \code{\link{tran.1D}} function.
}
\description{
This function sets up the diffusion advection model
}
