% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSinePop.R
\name{getOptimalPop}
\alias{getOptimalPop}
\title{Get Optimal Pop
Initilaizes optimal initial population distribution based on resource}
\usage{
getOptimalPop(tau, X.min = 0, X.max = 100, dx, x.peak, t.peak, x.sd, t.sd)
}
\arguments{
\item{tau}{maximum time in days for migratory population in one year}

\item{X.min}{minimum value of population distribution}

\item{X.max}{maximum value of population distribution}

\item{dx}{time step}

\item{x.peak}{where the resource peak is in space}

\item{t.peak}{where the resource peak is in time}

\item{x.sd}{standard deviation of resource peak in space}

\item{t.sd}{standard deviation of resource peak in time}
}
\description{
Get Optimal Pop
Initilaizes optimal initial population distribution based on resource
}
