% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateScripts.R
\name{parameterGrid}
\alias{parameterGrid}
\title{Parameter Grid}
\usage{
parameterGrid(epsilons, alphas, beta0s, beta1s)
}
\arguments{
\item{epsilons}{values of epsilon parameter}

\item{alphas}{values value of alpha parameter}

\item{beta0s}{values value of beta0 parameter}

\item{beta1s}{values value of beta1 parameter}
}
\value{
list of data frames
}
\description{
Generates a data frame splitting up all of the combinations of parameters
}
\examples{
parameterGrid(epsilon = 5, depsilon = 1, alpha = 5, dalpha = 1, beta0 = 3, dbeta0 = 1, beta1 = 2, dbeta1 = 0)
}
\seealso{
\link{createShellScript}, \link{createSource}, \link{createFinalShellScript}
}
