% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateScripts.R
\name{parameterGrid}
\alias{parameterGrid}
\title{Parameter Grid}
\usage{
parameterGrid(epsilons, alphas, beta0s, beta1s, existing = NULL)
}
\arguments{
\item{epsilons}{values of epsilon parameter}

\item{alphas}{values value of alpha parameter}

\item{beta0s}{values value of beta0 parameter}

\item{beta1s}{values value of beta1 parameter}

\item{existing}{a data frame with existing values already tested}
}
\value{
list of data frames
}
\description{
Generates a data frame splitting up all of the combinations of parameters
}
\examples{
parameterGrid(c(1,2,3), c(2,3), seq(1,4), 4)
}
\seealso{
\link{createShellScript}, \link{createSource}, \link{createFinalShellScript}
}
