% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotManyRuns}
\alias{plotManyRuns}
\title{Plotting simulation results}
\usage{
plotManyRuns(
  sim,
  years = NULL,
  nrow = 1,
  outer = TRUE,
  labelyears = FALSE,
  par = NULL,
  ylab = "",
  ...
)
}
\description{
Plotting simulation results
}
