% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotMigration}
\alias{plotMigration}
\title{Plot migration patterns over resource}
\usage{
plotMigration(M, world, plotresource = TRUE, add = FALSE)
}
\description{
Plot migration patterns over resource
}
