% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotYearList}
\alias{plotYearList}
\title{Plot Year List}
\usage{
plotYearList(
  yearlist,
  world,
  X.max = 100,
  tau = 360,
  log = FALSE,
  persp = FALSE,
  ...
)
}
\arguments{
\item{yearlist}{a list of T x X matrices describing the population 
distribution for each year after initial population; produced by the runManyYear function}

\item{X.max}{maximum value of X-axis}

\item{tau}{maximum value of time T}

\item{log}{logical;}

\item{persp}{whether to plot an image / contour or perspective plot}

\item{World}{List of 5: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau,
the dx value and the tau value. Can incorporate resource attribute into the world to make 
a list of 6.  Set up by the getSinePop function.}
}
\description{
This function plots the migratoriness population distribution over a time period.
}
\examples{
world <- getOptimalPop(tau=100, X.min = -100, X.max = 100, dx=1, 
                       x1 = 30, x2 = -30, t.peak=25, 
                       x.sd=12, t.sd=6)
world$m0 <- fitMigration(t = world$time, x = getMem(world$pop, world))

par0 <- getCCpars(mu_x0 = 30, 
                  mu_t0 = 25,
                  beta_x = 0,
                  beta_t = 0,
                  n.years = 30,
                  sigma_x = 12,
                  sigma_t = 6,
                  psi_x = 0, 
                  psi_t = 0,
                  n.years.null = 0) 

world$resource <- aaply(par0, 1, function(p) getResource_island(world, p))
attr(world$resource, "par") <- par0[nrow(par0),]

#run simulation for 12 years 

parameters <- c(epsilon = 4, alpha = 100, beta=400, kappa = 0, lambda = 80)

sim <- runManyYears(world, parameters = parameters, n.years = 12, threshold = 0.9999)


plotManyRuns(sim$pop, world)
plotYearList(sim$pop, world, tau = tau)
plotMemories(sim$pop, world)


}
\seealso{
\link{runManyYears}; \link{printParameters}
}
