% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotYearList}
\alias{plotYearList}
\title{Plot Year List}
\usage{
plotYearList(
  yearlist,
  world,
  X.max = 100,
  tau = 360,
  log = FALSE,
  persp = FALSE,
  ...
)
}
\arguments{
\item{yearlist}{a list of T x X matrices describing the population 
distribution for each year after initial population; produced by the runManyYear function}

\item{X.max}{maximum value of X-axis}

\item{tau}{maximum value of time T}

\item{log}{logical;}

\item{persp}{whether to plot an image / contour or perspective plot}

\item{World}{List of 5: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau,
the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 6.
Set up by the getSinePop function.}
}
\description{
This function plots the migratoriness population distribution over a time period.
}
\examples{
world.R1 <- getOptimalPop(tau = 100, t.peak = 25, 
                          x1 = 40, x2 = -40, x.sd = 5, t.sd = 12)
world.R1$resource <- getResource_island(world.R1, 
                                        c(t.peak = 25, t.sd = 12, x.peak = 35, x.sd = 5))

#run simulation for 12 years 

parameters <- c(epsilon = 5, alpha = 100, beta=50, kappa = 1, lambda = 20)

M <- runManyYears(world.R1, parameters = parameters, n.years = 12, threshold = 0.9999)

#simulation runs for 3 years since that is when threshold is hit so only 3 years are plotted
plotManyRuns(M, world.R1)
plotYearList(M, world.R1, tau = tau)
plotMemories(M, world.R1)

}
\seealso{
\link{runManyYears}; \link{printParameters}
}
