% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotYearList}
\alias{plotYearList}
\title{Plot Year List}
\usage{
plotYearList(yearlist, X.max = 100, tau = 360, log = FALSE, persp = FALSE, ...)
}
\arguments{
\item{yearlist}{a list of T x X matrices describing the population 
distribution for each year after initial population; produced by the runManyYear function}

\item{X.max}{maximum value of X-axis}

\item{tau}{maximum value of time T}

\item{log}{logical;}

\item{persp}{logical;}
}
\description{
This function plots the migratoriness population distribution over a time period.
}
\seealso{
\link{runManyYears}; \link{printParameters}
}
