% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotYearList}
\alias{plotYearList}
\title{Plot Year List}
\usage{
plotYearList(
  yearlist,
  world,
  X.max = 100,
  tau = 360,
  log = FALSE,
  persp = FALSE,
  ...
)
}
\arguments{
\item{yearlist}{a list of T x X matrices describing the population 
distribution for each year after initial population; produced by the runManyYear function}

\item{X.max}{maximum value of X-axis}

\item{tau}{maximum value of time T}

\item{log}{logical;}

\item{persp}{whether to plot an image / contour or perspective plot}

\item{World}{List of 5: a population distribution across the time period in a T x X matrix,
a vector with midpoint X-values, the time points for the population as integers 1:tau,
the dx value and the tau value. Can incorporate resource attribute into the world to make a list of 6.
Set up by the getSinePop function.}
}
\description{
This function plots the migratoriness population distribution over a time period.
}
\examples{
world <- getSinePop(tau = 100, X.min = 0, X.max = 100, dx=1, 
                    peak.max = 80, peak.min = 20, sd = 10)
par(mfrow = c(1,2))
plotYearList(list(world$pop))
plotYearList(list(world$pop), persp = TRUE)
}
\seealso{
\link{runManyYears}; \link{printParameters}
}
