% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModel.R
\name{runManyRuns}
\alias{runManyRuns}
\title{Run Many Years for a set of parameters}
\usage{
runManyRuns(parameters.df, world, ...)
}
\description{
Based on a migratory population's set up (the World) and the values \code{alpha},
\code{beta0} and \code{beta1}, this function determines the population 
distribution after several years.
}
