% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModel.R
\name{runManyRuns_res}
\alias{runManyRuns_res}
\title{Run Many Runs Resource}
\usage{
runManyRuns_res(
  world_param,
  parameters.df,
  resource_param,
  world,
  resource,
  filename = NULL,
  results.dir = NULL,
  ...
)
}
\description{
Running on server for optimal population to match resource
}
